// reelease-ai-next/src/types/components/features.ts
export type FeatureTag = {
  label: string
  icon: LucideIcon
  iconClassName?: string
}